+++ package.json
@@ -8,17 +8,20 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@react-native-community/masked-view": "0.1.10",
+    "@react-navigation/native": "^5.7.0",
+    "@react-navigation/stack": "^5.7.0",
+    "base-64": "^0.1.0",
     "expo": "~38.0.8",
     "expo-status-bar": "^1.0.2",
     "react": "~16.11.0",
     "react-dom": "~16.11.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-38.0.2.tar.gz",
-    "react-native-web": "~0.11.7",
     "react-native-gesture-handler": "~1.6.0",
     "react-native-reanimated": "~1.9.0",
-    "react-native-screens": "~2.9.0",
     "react-native-safe-area-context": "~3.0.7",
-    "@react-native-community/masked-view": "0.1.10"
+    "react-native-screens": "~2.9.0",
+    "react-native-web": "~0.11.7"
   },
   "devDependencies": {
     "@babel/core": "^7.8.6",
