+++ src/screens/ServiceNowForm.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react'
 let base64 = require('base-64')
-import {config} from '../credential/env';
+import { config } from '../credentials/env';
 import { View, Text, TouchableOpacity, TextInput, StyleSheet, Alert } from 'react-native'
 
-class ServiseNowForm extends Component {
+class ServiceNowForm extends Component {
    state = {
       Description: ''
    }
@@ -15,21 +15,17 @@ class ServiseNowForm extends Component {
       let username = config.USERNAME;
       let password = config.PASSWORD;
       let headers = new Headers();
-    
       let data = {
         'description': Description
       }
-      console.log(data)
-  
-      //headers.append('Content-Type', 'text/json');
       headers.append('Authorization', 'Basic ' + base64.encode(username + ":" + password));
       headers.append('Content-Type', 'application/json;charset=UTF-8') 
       fetch(url, {method:'POST',
          headers: headers,
          body: JSON.stringify(data)
-   })
+      })
       .then(response => {
-           console.log(response.status)
+         console.log(response.status)
          if(response.status==201)
         {
             Alert.alert("Ticket created "+Description)
@@ -47,11 +43,9 @@ class ServiseNowForm extends Component {
                underlineColorAndroid = "transparent"
                placeholder = "Write The Description Of Your Tickets"
                placeholderTextColor = "#FFFFFF"
-               multiline={true}
-               
+               multiline={true}      
                autoCapitalize = "none"
                onChangeText = {this.handleDescription}/>
-            
             <TouchableOpacity
                style = {styles.button}
                onPress = {
@@ -63,7 +57,7 @@ class ServiseNowForm extends Component {
       )
    }
 }
-export default ServiseNowForm
+export default ServiceNowForm
 
 const styles = StyleSheet.create({
    container: {
